worker/requirements.txt
httpx==0.27.2
APScheduler==3.10.4
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
pydantic==2.9.2
python-dotenv==1.0.1
orjson==3.10.7
python-dateutil==2.8.2